updated: applied black to all files

+++ backend/apps/channel_admins/signals.py
@@ -21,8 +21,8 @@ def send_admin_promotion_email(sender, created, instance, **kwargs):
             body={
                 "channel_title": channel.title,
                 "first_name": user.first_name,
-                "channel_token": channel.token
-            }
+                "channel_token": channel.token,
+            },
         )
 
 
@@ -32,8 +32,7 @@ def create_admin_after_creating_channel(sender, created, instance, **kwargs):
     """
 
     if created:
-
-        ChannelAdmin .objects.create(user=instance.owner, channel=instance)
+        ChannelAdmin.objects.create(user=instance.owner, channel=instance)
 
 
 @receiver(post_delete, sender=ChannelSubscriber)
@@ -42,18 +41,14 @@ def delete_admin_after_unsubscribing(sender, **kwargs):
     Delete admin after unsubscribing from channel.
     """
     instance = kwargs["instance"]
-    ChannelAdmin.objects.filter(
-        user=instance.user, channel=instance.channel
-    ).delete()
+    ChannelAdmin.objects.filter(user=instance.user, channel=instance.channel).delete()
 
 
-@receiver(post_save, sender=ChannelAdmin )
+@receiver(post_save, sender=ChannelAdmin)
 def create_permissions_for_admin(sender, instance, created, *args, **kwargs):
     """
     Create permission object after an admin is created.
     """
 
     if created:
-        ChannelAdminPermission.objects.create(
-            admin=instance
-        )
+        ChannelAdminPermission.objects.create(admin=instance)
